move most functionality into ark_curve
