+++ src/ark_curve/encoding.rs
@@ -5,10 +5,12 @@ use core::convert::{TryFrom, TryInto};
 use ark_ec::twisted_edwards::TECurveConfig;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, Read, Write};
 
-use crate::{
-    constants::TWO, element::Decaf377EdwardsConfig, sign::Sign, EdwardsProjective, Element,
-    EncodingError, Fq, OnCurve,
+use crate::ark_curve::{
+    constants::TWO, edwards::Decaf377EdwardsConfig, on_curve::OnCurve, EdwardsProjective, Element,
+    EncodingError,
 };
+use crate::sign::Sign;
+use crate::Fq;
 
 #[derive(Copy, Clone, Default, Eq, Ord, PartialOrd, PartialEq)]
 pub struct Encoding(pub [u8; 32]);
