+++ src/ark_curve/mod.rs
@@ -1,2 +1,39 @@
+pub mod bls12_377;
 mod constants;
 mod edwards;
+
+mod element;
+mod elligator;
+mod encoding;
+mod error;
+mod field_ext;
+mod invsqrt;
+mod ops;
+pub mod rand;
+pub mod serialize;
+
+pub use constants::ZETA;
+pub(crate) use edwards::{Decaf377EdwardsConfig, EdwardsProjective};
+pub use element::{AffineElement, Element};
+pub use encoding::Encoding;
+pub use error::EncodingError;
+pub use field_ext::FieldExt;
+
+mod on_curve;
+
+#[cfg(feature = "r1cs")]
+pub mod r1cs;
+
+pub use bls12_377::Bls12_377;
+
+/// Return the conventional generator for `decaf377`.
+pub fn basepoint() -> Element {
+    Element {
+        inner: EdwardsProjective::new(
+            *constants::B_X,
+            *constants::B_Y,
+            *constants::B_T,
+            *constants::B_Z,
+        ),
+    }
+}
